# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QProcure Quantum Decision Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Supplier {
  id: number;
  name: string;
  cost: number;
  riskScore: number;
  deliveryTime: number;
  createdAt: string;
}

export interface CreateSupplierBody {
  name: string;
  cost: number;
  riskScore: number;
  deliveryTime: number;
}

export interface InventoryItem {
  id: number;
  name: string;
  quantity: number;
  cost: number;
  lowStockThreshold: number;
  createdAt: string;
}

export interface CreateInventoryItemBody {
  name: string;
  quantity: number;
  cost: number;
  lowStockThreshold: number;
}

export interface SaleRecord {
  id: number;
  product: string;
  quantity: number;
  revenue: number;
  createdAt: string;
}

export interface CreateSaleBody {
  product: string;
  quantity: number;
  revenue: number;
}

export interface QuantumOptimizeRequest {
  supplierIds: number[];
  budget: number;
  algorithm?: string;
}

export interface QuantumMetrics {
  qubits: number;
  depth: number;
  shots: number;
  backend: string;
  executionTime: string;
}

export interface SelectedSupplier {
  id: number;
  name: string;
  cost: number;
  riskScore: number;
  deliveryTime: number;
}

export interface ProbabilityEntry {
  bitstring: string;
  probability: number;
}

export interface ConvergencePoint {
  iteration: number;
  cost: number;
}

export interface QuantumOptimizeResult {
  algorithmUsed: string;
  algorithmReason: string;
  bestSolution: string;
  selectedSuppliers: SelectedSupplier[];
  totalCost: number;
  classicalCost: number;
  improvement: string;
  confidence: number;
  probabilities: ProbabilityEntry[];
  convergence: ConvergencePoint[];
  quantumMetrics: QuantumMetrics;
  codeSnippet: string;
  circuitDiagram: string;
  statevectorAmplitudes: number[];
}

export interface QuantumStatus {
  available: boolean;
  backend: string;
  version: string;
  message: string;
}

export interface DashboardSummary {
  totalInventoryValue: number;
  activeSuppliers: number;
  salesRevenue: number;
  optimizationSavings: number;
  inventoryItems: number;
  totalSales: number;
  lowStockItems: number;
}

export interface TrendPoint {
  date: string;
  value: number;
  label: string;
}

export interface DashboardTrends {
  inventoryTrend: TrendPoint[];
  salesTrend: TrendPoint[];
  costComparison: TrendPoint[];
}