# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QProcure Quantum Decision Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all suppliers
 */
export const GetSuppliersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  cost: zod.number(),
  riskScore: zod.number(),
  deliveryTime: zod.number(),
  createdAt: zod.string(),
});
export const GetSuppliersResponse = zod.array(GetSuppliersResponseItem);

/**
 * @summary Create a supplier
 */
export const CreateSupplierBody = zod.object({
  name: zod.string(),
  cost: zod.number(),
  riskScore: zod.number(),
  deliveryTime: zod.number(),
});

/**
 * @summary Get a supplier by ID
 */
export const GetSupplierParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSupplierResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  cost: zod.number(),
  riskScore: zod.number(),
  deliveryTime: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Update a supplier
 */
export const UpdateSupplierParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSupplierBody = zod.object({
  name: zod.string(),
  cost: zod.number(),
  riskScore: zod.number(),
  deliveryTime: zod.number(),
});

export const UpdateSupplierResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  cost: zod.number(),
  riskScore: zod.number(),
  deliveryTime: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a supplier
 */
export const DeleteSupplierParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List inventory items
 */
export const GetInventoryResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  quantity: zod.number(),
  cost: zod.number(),
  lowStockThreshold: zod.number(),
  createdAt: zod.string(),
});
export const GetInventoryResponse = zod.array(GetInventoryResponseItem);

/**
 * @summary Create an inventory item
 */
export const CreateInventoryItemBody = zod.object({
  name: zod.string(),
  quantity: zod.number(),
  cost: zod.number(),
  lowStockThreshold: zod.number(),
});

/**
 * @summary Update an inventory item
 */
export const UpdateInventoryItemParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateInventoryItemBody = zod.object({
  name: zod.string(),
  quantity: zod.number(),
  cost: zod.number(),
  lowStockThreshold: zod.number(),
});

export const UpdateInventoryItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  quantity: zod.number(),
  cost: zod.number(),
  lowStockThreshold: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete an inventory item
 */
export const DeleteInventoryItemParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all sales records
 */
export const GetSalesResponseItem = zod.object({
  id: zod.number(),
  product: zod.string(),
  quantity: zod.number(),
  revenue: zod.number(),
  createdAt: zod.string(),
});
export const GetSalesResponse = zod.array(GetSalesResponseItem);

/**
 * @summary Create a sales record
 */
export const CreateSaleBody = zod.object({
  product: zod.string(),
  quantity: zod.number(),
  revenue: zod.number(),
});

/**
 * @summary Delete a sale record
 */
export const DeleteSaleParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Run quantum optimization algorithm
 */
export const RunQuantumOptimizationBody = zod.object({
  supplierIds: zod.array(zod.number()),
  budget: zod.number(),
  algorithm: zod.string().optional(),
});

export const RunQuantumOptimizationResponse = zod.object({
  algorithmUsed: zod.string(),
  algorithmReason: zod.string(),
  bestSolution: zod.string(),
  selectedSuppliers: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      cost: zod.number(),
      riskScore: zod.number(),
      deliveryTime: zod.number(),
    }),
  ),
  totalCost: zod.number(),
  classicalCost: zod.number(),
  improvement: zod.string(),
  confidence: zod.number(),
  probabilities: zod.array(
    zod.object({
      bitstring: zod.string(),
      probability: zod.number(),
    }),
  ),
  convergence: zod.array(
    zod.object({
      iteration: zod.number(),
      cost: zod.number(),
    }),
  ),
  quantumMetrics: zod.object({
    qubits: zod.number(),
    depth: zod.number(),
    shots: zod.number(),
    backend: zod.string(),
    executionTime: zod.string(),
  }),
  codeSnippet: zod.string(),
  circuitDiagram: zod.string(),
  statevectorAmplitudes: zod.array(zod.number()),
});

/**
 * @summary Get quantum backend status
 */
export const GetQuantumStatusResponse = zod.object({
  available: zod.boolean(),
  backend: zod.string(),
  version: zod.string(),
  message: zod.string(),
});

/**
 * @summary Get dashboard KPI summary
 */
export const GetDashboardSummaryResponse = zod.object({
  totalInventoryValue: zod.number(),
  activeSuppliers: zod.number(),
  salesRevenue: zod.number(),
  optimizationSavings: zod.number(),
  inventoryItems: zod.number(),
  totalSales: zod.number(),
  lowStockItems: zod.number(),
});

/**
 * @summary Get time-series trends for dashboard charts
 */
export const GetDashboardTrendsResponse = zod.object({
  inventoryTrend: zod.array(
    zod.object({
      date: zod.string(),
      value: zod.number(),
      label: zod.string(),
    }),
  ),
  salesTrend: zod.array(
    zod.object({
      date: zod.string(),
      value: zod.number(),
      label: zod.string(),
    }),
  ),
  costComparison: zod.array(
    zod.object({
      date: zod.string(),
      value: zod.number(),
      label: zod.string(),
    }),
  ),
});