# lib/api-zod/src/generated/types/createSaleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QProcure Quantum Decision Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSaleBody {
  product: string;
  quantity: number;
  revenue: number;
}