# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QProcure Quantum Decision Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalInventoryValue: number;
  activeSuppliers: number;
  salesRevenue: number;
  optimizationSavings: number;
  inventoryItems: number;
  totalSales: number;
  lowStockItems: number;
}